requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dxcam==0.0.5
ndi-python==5.1.1.5
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
pygame==2.5.0
PySimpleGUI==4.60.5
pystray==0.19.4
